etl/requirements.txt
# ETL Dependencies
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=0.8.11
xlrd>=2.0.1
numpy>=1.24.0
python-slugify>=8.0.0
chardet>=5.0.0

# Backend Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
